Alter the way in which R returns a file link to graphical output.
We currently have a (convenient) hack in R that, when :file filename
is used, automatically sends standard graphical output to
filename. We're planning some changes to the behavior of the :results
file and :file headers, and I think this R hack will be a
casualty (perhaps the behavior can live on with ":results graphics
file"). In any case, this change moves to a more conservative method
of returning file results that, I think, we can be sure will work in
the future. This requires org commit 14f4dbd.